backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
groq
pydantic
diskcache
jinja2
python-multipart
